requirements.txt
streamlit
requests
python-dotenv
folium
streamlit-folium
plotly
pandas
